docs+refactor: resolve architecture review nits from code-mode gate
INV-17: add MigrateEntityComponentsAsync step to flow-01-server-startup.md
  (mermaid + step 7 sub-step h/i) and flow-05-content-reload.md
  (mermaid + new step 10; existing 10-12 renumbered to 11-13).

INV-16/INV-D3: docs/reference/archetypes.md — replace "not yet built"
  banner with implementation-status note; update table to show as-built
  required components for the five registered archetypes (✓ Mob, Player,
  Room, Area, StaticItem); planned-only rows marked "not yet in registry";
  step 5 in "How to add" notes missing test framework.

INV-16: docs/reference/systems.md — add IArchetypeRegistry/ArchetypeRegistry
  entry; extend WorldContentLoader entry with new dependencies and
  MigrateEntityComponentsAsync description.

INV-D3: docs/roadmap/completed/slice-9-combat.md — add validation-fix
  shipped-pieces table (archetype system, migration, counterattack,
  HP status, CombatRoundResult move); update code-mode gate note.

INV-6: move CombatRoundResult + CombatRoundOutcome from ICombatSystem.cs
  (Systems namespace) to Core/Modules/Combat/CombatRoundResult.cs (Combat
  namespace); remove now-redundant using from CombatRoundEvent.cs.
  C# namespace resolution makes the types visible in Events/ and Handlers/
  sub-namespaces without additional using directives.
