+++ Core/Modules/Combat/Systems/ICombatSystem.cs
@@ -1,18 +1,11 @@
 namespace Hedron.Core.Modules.Combat.Systems
 {
-    public readonly record struct CombatRoundResult(
-        uint AttackerEntityId,
-        uint DefenderEntityId,
-        int DamageDealt,
-        bool AttackerHit,
-        CombatRoundOutcome Outcome);
-
-    public enum CombatRoundOutcome { Hit, Miss, MobDied, PlayerIncapacitated }
-
     /// <summary>
     /// Domain system for combat resolution. Pure: no events, no persistence (INV-5, INV-8).
     /// Computes attack resolution via <c>IStatSystem</c> and mutates HP via
-    /// <c>IAttributeSystem.SetCurrentHp</c>. Returns structured results to callers.
+    /// <c>IAttributeSystem.SetCurrentHp</c>. Returns structured <see cref="CombatRoundResult"/>
+    /// to callers. Result types live in the parent <c>Combat</c> namespace so events and handlers
+    /// can reference them without importing <c>Systems</c>.
     /// </summary>
     public interface ICombatSystem
     {
